Restructured the Test Scripts
