+++ test/EPNS_StagingTests/EPNSCoreV1.Notifications.js
@@ -8,12 +8,12 @@ const {
   increase,
   increaseTo,
   latest,
-} = require("../../time");
-const { calcChannelFairShare, calcSubscriberFairShare, getPubKey, bn, tokens, tokensBN, bnToInt, ChannelAction, readjustFairShareOfChannels, SubscriberAction, readjustFairShareOfSubscribers } = require("../../../helpers/utils");
+} = require("../time");
+const { calcChannelFairShare, calcSubscriberFairShare, getPubKey, bn, tokens, tokensBN, bnToInt, ChannelAction, readjustFairShareOfChannels, SubscriberAction, readjustFairShareOfSubscribers } = require("../../helpers/utils");
 
 use(solidity);
 
-describe("EPNSStagingV1 tests", function () {
+describe("EPNSStagingV4 tests", function () {
   const AAVE_LENDING_POOL = "0x1c8756FD2B28e9426CDBDcC7E3c4d64fa9A54728";
   const DAI = "0xf80A32A835F79D7787E8a8ee5721D0fEaFd78108";
   const ADAI = "0xcB1Fe6F440c49E9290c3eb7f158534c2dC374201";
@@ -88,8 +88,8 @@ describe("EPNSStagingV1 tests", function () {
     const EPNSTOKEN = await ethers.getContractFactory("EPNS");
     EPNS = await EPNSTOKEN.deploy();
 
-    const EPNSStagingV1 = await ethers.getContractFactory("EPNSStagingV1");
-    LOGIC = await EPNSStagingV1.deploy();
+    const EPNSStagingV4 = await ethers.getContractFactory("EPNSStagingV4");
+    LOGIC = await EPNSStagingV4.deploy();
 
     const TimeLock = await ethers.getContractFactory("Timelock");
     TIMELOCK = await TimeLock.deploy(ADMIN, delay);
@@ -109,7 +109,7 @@ describe("EPNSStagingV1 tests", function () {
     );
 
     await EPNSProxy.changeAdmin(ALICESIGNER.address);
-    EPNSCoreV1Proxy = EPNSStagingV1.attach(EPNSProxy.address)
+    EPNSCoreV1Proxy = EPNSStagingV4.attach(EPNSProxy.address)
   });
 
   afterEach(function () {
@@ -126,11 +126,11 @@ describe("EPNSStagingV1 tests", function () {
         const CHANNEL_TYPE = 2;
         const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
 
-        await EPNSCoreV1Proxy.connect(ADMINSIGNER).addToChannelizationWhitelist(CHANNEL_CREATOR, {gasLimit: 500000});
+        //await EPNSCoreV1Proxy.connect(ADMINSIGNER).addToChannelizationWhitelist(CHANNEL_CREATOR, {gasLimit: 500000});
       
         await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
         await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
-        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel,{gasLimit: 500000});
+        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel,ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
       });
 
       it("should revert if anyone other than owner calls the function", async function(){
@@ -147,6 +147,58 @@ describe("EPNSStagingV1 tests", function () {
           .to.emit(EPNSCoreV1Proxy, 'SendNotification')
           .withArgs(CHANNEL_CREATOR, BOB, ethers.utils.hexlify(msg));
       });
+    });
+
+    describe("Testing sendNotificationAsDelegate function", function(){
+      beforeEach(async function(){
+        const CHANNEL_TYPE = 2;
+        const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
+      
+        await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
+        await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
+        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel,ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
+      });
+
+       it("No one except a Delegate should be able to send notification on behalf of a Channel", async function(){
+        const msg = ethers.utils.toUtf8Bytes("This is DELAGATED notification message");
+        const tx =  EPNSCoreV1Proxy.connect(BOBSIGNER).sendNotificationAsDelegate(CHANNEL_CREATOR,BOB,msg);
+        await expect(tx).to.be.revertedWith("Not authorised to send messages");
+      });
+
+      it("BOB Should be able to Send Delegated Notification once Allowed", async function(){
+        // Adding BOB As Delate Notification Seder
+        const tx_addDelegate =  await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).addDelegate(BOB);
+        const isBobAllowed = await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).delegated_NotificationSenders(CHANNEL_CREATOR,BOB);
+        
+        // BOB Sending Delegated Notification
+        const msg = ethers.utils.toUtf8Bytes("This is DELAGATED notification message");
+        const tx_sendNotif =  await EPNSCoreV1Proxy.connect(BOBSIGNER).sendNotificationAsDelegate(CHANNEL_CREATOR,ALICE,msg);
+        
+        await expect(tx_sendNotif)
+          .to.emit(EPNSCoreV1Proxy, 'SendNotification')
+          .withArgs(CHANNEL_CREATOR, ALICE, ethers.utils.hexlify(msg));
+        await expect(isBobAllowed).to.be.equal(true);
+        await expect(tx_addDelegate)
+          .to.emit(EPNSCoreV1Proxy, 'AddDelegate')
+          .withArgs(CHANNEL_CREATOR, BOB);
+      })
+      
+       it("BOB Should NOT be able to Send Delegated Notification once Permission is Revoked", async function(){
+        // Revoking Permission from BOB
+        const tx_removeDelegate =  EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).removeDelegate(BOB);
+        const isBobAllowed = await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).delegated_NotificationSenders(CHANNEL_CREATOR,BOB);
+
+        // BOB Sending Delegated Notification
+         const msg = ethers.utils.toUtf8Bytes("This is DELAGATED notification message");
+        const tx_sendNotif =  EPNSCoreV1Proxy.connect(BOBSIGNER).sendNotificationAsDelegate(CHANNEL_CREATOR,BOB,msg);
+        
+
+        await expect(tx_sendNotif).to.be.revertedWith("Not authorised to send messages");
+        await expect(isBobAllowed).to.be.equal(false);
+          await expect(tx_removeDelegate)
+          .to.emit(EPNSCoreV1Proxy, 'RemoveDelegate')
+          .withArgs(CHANNEL_CREATOR, BOB);
+      })
     });
 
   });
