+++ test/EPNS_StagingTests/EPNSCoreV1.Subscribe.js
@@ -8,12 +8,12 @@ const {
   increase,
   increaseTo,
   latest,
-} = require("../../time");
-const { calcChannelFairShare, calcSubscriberFairShare, getPubKey, bn, tokens, tokensBN, bnToInt, ChannelAction, readjustFairShareOfChannels, SubscriberAction, readjustFairShareOfSubscribers } = require("../../../helpers/utils");
+} = require("../time");
+const { calcChannelFairShare, calcSubscriberFairShare, getPubKey, bn, tokens, tokensBN, bnToInt, ChannelAction, readjustFairShareOfChannels, SubscriberAction, readjustFairShareOfSubscribers } = require("../../helpers/utils");
 
 use(solidity);
 
-describe("EPNSStagingV1 tests", function () {
+describe("EPNSStagingV4 tests", function () {
   const AAVE_LENDING_POOL = "0x1c8756FD2B28e9426CDBDcC7E3c4d64fa9A54728";
   const DAI = "0xf80A32A835F79D7787E8a8ee5721D0fEaFd78108";
   const ADAI = "0xcB1Fe6F440c49E9290c3eb7f158534c2dC374201";
@@ -88,8 +88,8 @@ describe("EPNSStagingV1 tests", function () {
     const EPNSTOKEN = await ethers.getContractFactory("EPNS");
     EPNS = await EPNSTOKEN.deploy();
 
-    const EPNSStagingV1 = await ethers.getContractFactory("EPNSStagingV1");
-    LOGIC = await EPNSStagingV1.deploy();
+    const EPNSStagingV4 = await ethers.getContractFactory("EPNSStagingV4");
+    LOGIC = await EPNSStagingV4.deploy();
 
     const TimeLock = await ethers.getContractFactory("Timelock");
     TIMELOCK = await TimeLock.deploy(ADMIN, delay);
@@ -109,7 +109,7 @@ describe("EPNSStagingV1 tests", function () {
     );
 
     await EPNSProxy.changeAdmin(ALICESIGNER.address);
-    EPNSCoreV1Proxy = EPNSStagingV1.attach(EPNSProxy.address)
+    EPNSCoreV1Proxy = EPNSStagingV4.attach(EPNSProxy.address)
   });
 
   afterEach(function () {
@@ -119,8 +119,9 @@ describe("EPNSStagingV1 tests", function () {
     EPNSProxy = null
     EPNSCoreV1Proxy = null
   });
-  
-  describe("Testing subscribe realted functions", function(){
+
+ 
+ describe("Testing subscribe realted functions", function(){
         /**
      * "subscribe" Function CHECKPOINTS
      *  Should only be called for Activated Channels
@@ -131,105 +132,105 @@ describe("EPNSStagingV1 tests", function () {
      * Should update the FAIRSHARE COUNTS
     //  */
 
-    describe("Testing the Base SUBSCRIBE Function", function(){ 
-      const CHANNEL_TYPE = 2;
-      const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
+  //   describe("Testing the Base SUBSCRIBE Function", function(){ 
+  //     const CHANNEL_TYPE = 2;
+  //     const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
   
-      beforeEach(async function(){
-     const CHANNEL_TYPE = 2;
-        const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
+  //     beforeEach(async function(){
+     
+  //       await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
+  //       await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
+  //       await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
 
-        await EPNSCoreV1Proxy.connect(ADMINSIGNER).addToChannelizationWhitelist(CHANNEL_CREATOR, {gasLimit: 500000});
-      
-        await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
-        await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
-        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel,{gasLimit: 500000});
-      })
+  //       await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(DELEGATED_CONTRACT_FEES);
+  //       await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, DELEGATED_CONTRACT_FEES);
+  //     })
 
-      it("User should be able to subscribe ONLY if Channel Is Activated", async ()=>{
-          const tx = EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(ALICE);
-          await expect(tx).to.be.revertedWith("Channel deactivated or doesn't exists");
-      })
+  //     it("User should be able to subscribe ONLY if Channel Is Activated", async ()=>{
+  //         const tx = EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(ALICE);
+  //         await expect(tx).to.be.revertedWith("Channel deactivated or doesn't exists");
+  //     })
 
-       it("Function should revert if user already subscribed", async function () {
-        await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
-        const tx = EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).subscribe(CHANNEL_CREATOR);
+  //      it("Function should revert if user already subscribed", async function () {
+  //       await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
+  //       const tx = EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).subscribe(CHANNEL_CREATOR);
         
-        await expect(tx).to.be.revertedWith("Subscriber already Exists");
-      });
+  //       await expect(tx).to.be.revertedWith("Subscriber already Exists");
+  //     });
 
-      it("Should Add User in the Ecosystem if they aren't added already", async function () {
-          const userDetails = await EPNSCoreV1Proxy.users(BOB);
-          const userActivatedBefore = userDetails[0];
-          const totalUsers_before = await EPNSCoreV1Proxy.usersCount();
+  //     it("Should Add User in the Ecosystem if they aren't added already", async function () {
+  //         const userDetails = await EPNSCoreV1Proxy.users(BOB);
+  //         const userActivatedBefore = userDetails[0];
+  //         const totalUsers_before = await EPNSCoreV1Proxy.usersCount();
 
-          await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
-          const userDetails_after = await EPNSCoreV1Proxy.users(BOB);
-          const userActivatedAfter= userDetails_after[0];
-          const totalUsers_after =  await EPNSCoreV1Proxy.usersCount();
+  //         await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
+  //         const userDetails_after = await EPNSCoreV1Proxy.users(BOB);
+  //         const userActivatedAfter= userDetails_after[0];
+  //         const totalUsers_after =  await EPNSCoreV1Proxy.usersCount();
    
 
-          expect(userActivatedBefore).to.be.equals(false);
-          expect(userActivatedAfter).to.be.equals(true);
-          expect(totalUsers_before).to.equal(3);
-          expect(totalUsers_after).to.equal(4);
-      });
+  //         expect(userActivatedBefore).to.be.equals(false);
+  //         expect(userActivatedAfter).to.be.equals(true);
+  //         expect(totalUsers_before).to.equal(3);
+  //         expect(totalUsers_after).to.equal(4);
+  //     });
 
-      it("Should Update the User & Channel Subscription Details on the Contract", async ()=>{
-          const userDetails = await EPNSCoreV1Proxy.users(BOB);
-          const channelDetails = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
+  //     it("Should Update the User & Channel Subscription Details on the Contract", async ()=>{
+  //         const userDetails = await EPNSCoreV1Proxy.users(BOB);
+  //         const channelDetails = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
           
-          const userSubscribeCount_before = userDetails[4];
-          const channelMemberCount_before = channelDetails[3];
+  //         const userSubscribeCount_before = userDetails[4];
+  //         const channelMemberCount_before = channelDetails[3];
 
-          const tx =  await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
+  //         const tx =  await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
 
-          const userDetails_after = await EPNSCoreV1Proxy.users(BOB);
-          const channelDetails_after = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
-          const channelMemberCount_after = channelDetails_after[3];
+  //         const userDetails_after = await EPNSCoreV1Proxy.users(BOB);
+  //         const channelDetails_after = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
+  //         const channelMemberCount_after = channelDetails_after[3];
 
-          const userSubscribeCount_after = userDetails_after[4];
-          const memberExists = await EPNSCoreV1Proxy.memberExists(BOB,CHANNEL_CREATOR);
+  //         const userSubscribeCount_after = userDetails_after[4];
+  //         const memberExists = await EPNSCoreV1Proxy.memberExists(BOB,CHANNEL_CREATOR);
 
-          expect(memberExists).to.be.equals(true);
-          expect(userSubscribeCount_before).to.equal(0);
-          expect(userSubscribeCount_after).to.equal(1);
-          expect(channelMemberCount_before).to.equal(1);
-          expect(channelMemberCount_after).to.equal(2);
-          await expect(tx).to.emit(EPNSCoreV1Proxy,'Subscribe')
-          .withArgs(CHANNEL_CREATOR,BOB)
-      })
+  //         expect(memberExists).to.be.equals(true);
+  //         expect(userSubscribeCount_before).to.equal(0);
+  //         expect(userSubscribeCount_after).to.equal(1);
+  //         expect(channelMemberCount_before).to.equal(1);
+  //         expect(channelMemberCount_after).to.equal(2);
+  //         await expect(tx).to.emit(EPNSCoreV1Proxy,'Subscribe')
+  //         .withArgs(CHANNEL_CREATOR,BOB)
+  //     })
 
 
-      it("Should update the FAIR SHARE RATIO", async ()=>{
-        const channel = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
+  //     it("Should update the FAIR SHARE RATIO", async ()=>{
+  //       const channel = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
 
-        const _channelFairShareCount = channel.channelFairShareCount;
-        const _channelHistoricalZ = channel.channelHistoricalZ;
-        const _channelLastUpdate = channel.channelLastUpdate;
+  //       const _channelFairShareCount = channel.channelFairShareCount;
+  //       const _channelHistoricalZ = channel.channelHistoricalZ;
+  //       const _channelLastUpdate = channel.channelLastUpdate;
         
-        const tx = await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
-        const blockNumber = tx.blockNumber;
+  //       const tx = await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
+  //       const blockNumber = tx.blockNumber;
         
-        const { 
-          channelNewFairShareCount, 
-          channelNewHistoricalZ, 
-          channelNewLastUpdate, 
-        } = readjustFairShareOfSubscribers(SubscriberAction.SubscriberAdded, _channelFairShareCount, _channelHistoricalZ, _channelLastUpdate, bn(blockNumber));
+  //       const { 
+  //         channelNewFairShareCount, 
+  //         channelNewHistoricalZ, 
+  //         channelNewLastUpdate, 
+  //       } = readjustFairShareOfSubscribers(SubscriberAction.SubscriberAdded, _channelFairShareCount, _channelHistoricalZ, _channelLastUpdate, bn(blockNumber));
         
-        const channelNew = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
+  //       const channelNew = await EPNSCoreV1Proxy.channels(CHANNEL_CREATOR);
 
-        const _channelNewFairShareCountNew = channelNew.channelFairShareCount;
-        const _channelHistoricalZNew = channelNew.channelHistoricalZ;
-        const _channelLastUpdateNew = channelNew.channelLastUpdate;
+  //       const _channelNewFairShareCountNew = channelNew.channelFairShareCount;
+  //       const _channelHistoricalZNew = channelNew.channelHistoricalZ;
+  //       const _channelLastUpdateNew = channelNew.channelLastUpdate;
         
-        expect(_channelNewFairShareCountNew).to.equal(channelNewFairShareCount);
-        expect(_channelHistoricalZNew).to.equal(channelNewHistoricalZ);
-        expect(_channelLastUpdateNew).to.equal(channelNewLastUpdate);
-      })
+  //       expect(_channelNewFairShareCountNew).to.equal(channelNewFairShareCount);
+  //       expect(_channelHistoricalZNew).to.equal(channelNewHistoricalZ);
+  //       expect(_channelLastUpdateNew).to.equal(channelNewLastUpdate);
+  //     })
 
-  });
-  
+  // });
+
+
    /**
      * "unsubscribe" Function CHECKPOINTS
      * Should only be called for Valid Channels -> users[_channel].channellized == true
@@ -248,14 +249,13 @@ describe("EPNSStagingV1 tests", function () {
       const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
   
       beforeEach(async function(){
-      const CHANNEL_TYPE = 2;
-        const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
-
-        await EPNSCoreV1Proxy.connect(ADMINSIGNER).addToChannelizationWhitelist(CHANNEL_CREATOR, {gasLimit: 500000});
       
         await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
         await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
-        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel,{gasLimit: 500000});
+        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
+
+        await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(DELEGATED_CONTRACT_FEES);
+        await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, DELEGATED_CONTRACT_FEES);
       })
 
       it("Function should only be accessibly for VALID Channels", async()=>{
@@ -288,9 +288,9 @@ describe("EPNSStagingV1 tests", function () {
         await expect(tx)
           .to.emit(EPNSCoreV1Proxy, 'Unsubscribe')
           .withArgs(CHANNEL_CREATOR, BOB)
-      }).timeout(140000);;
+      });
 
-  it(" Should Update Imperative On-Chain Information for User",async()=>{
+    it(" Should Update Imperative On-Chain Information for User",async()=>{
         await EPNSCoreV1Proxy.connect(BOBSIGNER).subscribe(CHANNEL_CREATOR);
         
         // Contract State Before Unsubscribing
@@ -311,7 +311,7 @@ describe("EPNSStagingV1 tests", function () {
         expect(userAfter.subscribedCount).to.equal(userBefore.subscribedCount.sub(1))
         expect(channelAfter.memberCount).to.equal(channelBefore.memberCount.sub(1))      
 
-      }).timeout(200000);
+      });
 
       it("Should subscribe and update fair share values", async function(){
         const publicKey = await getPubKey(BOBSIGNER);
@@ -340,7 +340,7 @@ describe("EPNSStagingV1 tests", function () {
         expect(_channelNewFairShareCountNew).to.equal(channelNewFairShareCount);
         expect(_channelHistoricalZNew).to.equal(channelNewHistoricalZ);
         expect(_channelLastUpdateNew).to.equal(channelNewLastUpdate);
-      }).timeout(200000);
+      });
 
 
 
@@ -362,15 +362,10 @@ describe("EPNSStagingV1 tests", function () {
       const CHANNEL_TYPE = 2;
       const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
   
-      beforeEach(async function(){
-      const CHANNEL_TYPE = 2;
-        const testChannel = ethers.utils.toUtf8Bytes("test-channel-hello-world");
-
-        await EPNSCoreV1Proxy.connect(ADMINSIGNER).addToChannelizationWhitelist(CHANNEL_CREATOR, {gasLimit: 500000});
-      
+      beforeEach(async function(){      
         await MOCKDAI.connect(CHANNEL_CREATORSIGNER).mint(ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
         await MOCKDAI.connect(CHANNEL_CREATORSIGNER).approve(EPNSCoreV1Proxy.address, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
-        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel,{gasLimit: 500000});
+        await EPNSCoreV1Proxy.connect(CHANNEL_CREATORSIGNER).createChannelWithFees(CHANNEL_TYPE, testChannel, ADD_CHANNEL_MIN_POOL_CONTRIBUTION);
       })
   
       it("should revert subscribe if channels are deactivated", async function () {
@@ -493,7 +488,5 @@ describe("EPNSStagingV1 tests", function () {
       });
    
     });
-
-
-  });
+});
 });
